dhtread: use int for pnum

+++ Source/dthread.cpp
@@ -14,12 +14,12 @@
 namespace devilution {
 
 struct DThreadPkt {
-	uint8_t pnum;
+	int pnum;
 	_cmd_id cmd;
 	std::unique_ptr<byte[]> data;
 	uint32_t len;
 
-	DThreadPkt(uint8_t pnum, _cmd_id(cmd), std::unique_ptr<byte[]> data, uint32_t len)
+	DThreadPkt(int pnum, _cmd_id(cmd), std::unique_ptr<byte[]> data, uint32_t len)
 		: pnum(pnum)
 		, cmd(cmd)
 		, data(std::move(data))
@@ -80,7 +80,7 @@ void dthread_remove_player(uint8_t pnum)
 	sgMemCrit.Leave();
 }
 
-void dthread_send_delta(uint8_t pnum, _cmd_id cmd, std::unique_ptr<byte[]> data, uint32_t len)
+void dthread_send_delta(int pnum, _cmd_id cmd, std::unique_ptr<byte[]> data, uint32_t len)
 {
 	if (!gbIsMultiplayer)
 		return;
