+++ Source/dthread.h
@@ -11,7 +11,7 @@
 namespace devilution {
 
 void dthread_remove_player(uint8_t pnum);
-void dthread_send_delta(uint8_t pnum, _cmd_id cmd, std::unique_ptr<byte[]> data, uint32_t len);
+void dthread_send_delta(int pnum, _cmd_id cmd, std::unique_ptr<byte[]> data, uint32_t len);
 void dthread_start();
 void DThreadCleanup();
